Flux Chat: use folders structure similar to other examples
