+++ examples/flux-chat/js/ChatAppDispatcher.js
@@ -1,0 +1,5 @@
+import { Dispatcher } from 'flux';
+
+const ChatAppDispatcher = new Dispatcher();
+
+export default ChatAppDispatcher;
