+++ examples/flux-chat/js/actions/MessageActionTypes.js
