+++ examples/flux-chat/js/actions/ThreadActionTypes.js
