--- examples/flux-chat/js/core/AppDispatcher.js
@@ -1,5 +1,0 @@
-import { Dispatcher } from 'flux';
-
-const AppDispatcher = new Dispatcher();
-
-export default AppDispatcher;
