+++ examples/flux-chat/js/records/MessageRecord.js
