+++ examples/flux-chat/js/records/ThreadRecord.js
