+++ examples/flux-chat/js/stores/__tests__/MessageBufferStore-test.js
@@ -10,12 +10,19 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import Dispatcher from './AppDispatcher';
-import { List } from 'immutable';
-import { loadMessages } from '../message/MessageActions';
-import { loadThreads } from '../thread/ThreadActions';
+jest.disableAutomock();
 
-export function loadChatData(chatAppData) {
-  loadMessages(new List(chatAppData.messages));
-  loadThreads(new List(chatAppData.threads));
-}
+import MessageBufferStore from '../MessageBufferStore';
+import ActionTypes from '../../actions/MessageActionTypes';
+import { Map } from 'immutable';
+
+describe('MessageBufferStore', () => {
+  it('should clear the buffer when new message is created', () => {
+    const initialState = new Map({ t1: 'hello' });
+    const action = { type: ActionTypes.MESSAGE_CREATED, message: { threadId: 't1' } };
+
+    const newState = MessageBufferStore.reduce(initialState, action);
+
+    expect(newState.toJS()).toEqual({ t1: '' });
+  });
+});
