+++ examples/flux-chat/js/stores/__tests__/MessageListStore-test.js
